pyproject.toml
[project]
name = "mydjangoproject"
version = "0.1.0"
description = "Django REST API project"
requires-python = ">=3.12"
dependencies = [
    "Django>=5.1,<5.2",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.3.0",
    "python-dotenv>=1.0.0",
    "pillow>=12.0.0",
]
